Extract mapped elements from mapper component
